pyproject.toml
[project]
name = "illallangi-django-data"
authors = [
    { name = "Andrew Cole", email = "andrew.cole@illallangi.com" }
]
maintainers = [
    { name = "Andrew Cole", email = "andrew.cole@illallangi.com" }
]
description = "A Django application for Coley's Data Lake."
readme = "README.md"
requires-python = ">=3.10.1"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-createsuperuserwithpassword>=2.0.0",
    "django-extensions>=3.2.3",
    "django-htmx>=1.19.0",
    "django>=5.1.2",
    "djp",
    "python-dotenv>=1.0.1",
    "toposort>=1.10",
]
keywords = [
    "django",
    "data",
    "data-lake",
]
version = "0.0.0"
classifiers =[
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
]
license = { text = "MIT License" }
[project.urls]
Homepage = "https://github.com/illallangi/django-data-server"
Repository = "https://github.com/illallangi/django-data-server.git"
Issues = "https://github.com/illallangi/django-data-server/issues"
Changelog = "https://github.com/illallangi/django-data-server/blob/master/CHANGELOG.md"
[build-system]
requires = [
    "ruff>=0.6.9",
    "hatchling>=1.25.0",
]
build-backend = "hatchling.build"
[tool.deptry.per_rule_ignores]
DEP002 = [
    "django-createsuperuserwithpassword",
    "django-extensions",
    "django-htmx",
]
[tool.djlint]
profile="django"
format_css=true
format_js=true
ignore="H006"
[tool.ruff]
lint.select = ["ALL"]
lint.ignore = [    
    "COM812",
    "D",
    "E501",
    "ERA001",
    "ISC001",
]
exclude = [
    "**/migrations/0*.py",
]
[tool.ruff.lint.pydocstyle]
convention = "pep257"
[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"
[tool.uv]
dev-dependencies = [
    "commitizen>=3.29.1",
    "deptry>=0.20.0",
    "djlint>=1.35.2",
    "hatchling>=1.25.0",
    "pre-commit>=3.8.0",
    "ruff>=0.6.9",
]

[tool.uv.sources]
djp = { git = "https://github.com/andrewcole/djp" }
[tool.hatch.build.targets.sdist]
exclude = [
    "Makefile",
    "uv.lock",
]
[tool.hatch.build.targets.wheel]
packages = ["src/illallangi"]
exclude = [
    "Makefile",
    "uv.lock",
]
[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version",
    "src/illallangi/django/data/__version__.py:__version__",
]
pre_bump_hooks = [
  "uv sync --no-dev"
]
update_changelog_on_bump = true
major_version_zero = true
